feat: extract mesh builder into module
